Fixed invalid types

+++ lib/MC/MCAsmStreamer.cpp
@@ -673,7 +673,7 @@ void MCAsmStreamer::EmitValueImpl(const MCExpr *Value, unsigned Size) {
     Directive = MAI->getData32bitsDirective();
     // If the target doesn't support 32-bit data, emit as two 16-bit halves.
     if (Directive) break;
-    int32_t Int32Value;
+    int64_t Int32Value;
     if (!Value->EvaluateAsAbsolute(Int32Value))
       report_fatal_error("Don't know how to emit this value.");
     if (MAI->isLittleEndian()) {
